fix: add version selection to all relevant commands
