+++ src/cmds/docs/index.js
@@ -7,6 +7,7 @@ const crypto = require('crypto');
 const frontMatter = require('gray-matter');
 const { promisify } = require('util');
 const APIError = require('../../lib/apiError');
+const { getSwaggerVersion } = require('../../lib/versionSelect');
 
 const readFile = promisify(fs.readFile);
 
@@ -42,13 +43,13 @@ exports.run = async function (opts) {
     return Promise.reject(new Error('No project API key provided. Please use `--key`.'));
   }
 
-  if (!version) {
-    return Promise.reject(new Error('No project version provided. Please use `--version`.'));
-  }
-
   if (!folder) {
     return Promise.reject(new Error(`No folder provided. Usage \`${config.cli} ${exports.usage}\`.`));
   }
+
+  const selectedVersion = await getSwaggerVersion(version, key).catch(e => {
+    return Promise.reject(e);
+  });
 
   // Find the files to sync
   const readdirRecursive = folderToSearch => {
@@ -72,7 +73,7 @@ exports.run = async function (opts) {
   const options = {
     auth: { user: key },
     headers: {
-      'x-readme-version': version,
+      'x-readme-version': selectedVersion,
     },
   };
